refactor(ui): CC-544 extract CustomSelect into a reusable component
Moves the labeled dropdown-select used by the JN drawer's Project/City
filters out of JNDrawer.tsx and into components/ui/custom-select.tsx,
matching where the other shared UI primitives (NavigationAccordion,
menu, drawer) already live, so it can be reused elsewhere without
duplicating it.
